Adds yarn-based (instead of npm) circleci CI/CD config
We use circle because it is very simple to setup and took less than 20
mins to implement and test.

Here we use a rolling docker image for node.js, so we get constantly
test against latest. At some point it may be prudent to schedule a build
so we can disambiguate between code-fails and node upgrade fails.